+++ src/main/java/org/example/services/impl/CategoryServiceImpl.java
@@ -1,6 +1,9 @@
 package org.example.services.impl;
 
 import org.example.dto.CategoryDto;
+import org.example.exception.CategoryAlreadyExistsException;
+import org.example.exception.CategoryNotFoundException;
+import org.example.exception.errorMessage.ErrorMessage;
 import org.example.mappers.CategoryMapper;
 import org.example.models.Category;
 import org.example.repositories.CategoryRepository;
@@ -10,7 +13,6 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -34,13 +36,19 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Optional<CategoryDto> getCategoryById(Long id) {
+    public CategoryDto getCategoryById(Long id) {
         return categoryRepository.findById(id)
-                .map(categoryMapper::toDto);
+                .map(categoryMapper::toDto)
+                .orElseThrow(() -> new CategoryNotFoundException(ErrorMessage.CATEGORY_NOT_FOUND));
     }
 
     @Override
     public CategoryDto createCategory(CategoryDto categoryDto) {
+        // Предполагаем, что категория уникальна по названию
+        categoryRepository.findByName(categoryDto.getName()).ifPresent(category -> {
+            throw new CategoryAlreadyExistsException(ErrorMessage.CATEGORY_ALREADY_EXISTS);
+        });
+
         Category category = categoryMapper.toEntity(categoryDto);
         category.setCreatedAt(LocalDateTime.now());
         Category savedCategory = categoryRepository.save(category);
@@ -48,17 +56,21 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Optional<CategoryDto> updateCategory(Long id, CategoryDto categoryDto) {
-        return categoryRepository.findById(id).map(category -> {
-            categoryMapper.updateEntityFromDto(categoryDto, category);
-            category.setUpdatedAt(LocalDateTime.now());
-            Category updatedCategory = categoryRepository.save(category);
-            return categoryMapper.toDto(updatedCategory);
-        });
+    public CategoryDto updateCategory(Long id, CategoryDto categoryDto) {
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() -> new CategoryNotFoundException(ErrorMessage.CATEGORY_NOT_FOUND));
+
+        categoryMapper.updateEntityFromDto(categoryDto, category);
+        category.setUpdatedAt(LocalDateTime.now());
+        Category updatedCategory = categoryRepository.save(category);
+        return categoryMapper.toDto(updatedCategory);
     }
 
     @Override
     public void deleteCategory(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new CategoryNotFoundException(ErrorMessage.CATEGORY_NOT_FOUND);
+        }
         categoryRepository.deleteById(id);
     }
 }
